backend/requirements.txt
python-dotenv
groq
yt-dlp
fastapi
uvicorn
weasyprint
markdown
pydantic